refactor: CloudPrivacyNotice in feature folder
